src/requirements.txt
nltk.download('words')